pyproject.toml
[project]
name = "dong-read"
version = "0.5.0"
description = "读咚咚 (Read DongDong) - 个人知识数据层的命令行接口"
readme = "readme.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "gudong"}
]
keywords = ["cli", "knowledge", "agent", "mcp", "read-it-later"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.12.0",
    "dong-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
]
mcp = [
    "mcp>=0.1.0",
]

[project.scripts]
dong-read = "read.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/read"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=read --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
